+++ lua/neotest-golang/runspec_test.lua
@@ -6,6 +6,7 @@ local convert = require("neotest-golang.convert")
 local options = require("neotest-golang.options")
 local json = require("neotest-golang.json")
 local cmd = require("neotest-golang.cmd")
+local dap = require("neotest-golang.dap")
 
 local M = {}
 
@@ -31,8 +32,8 @@ function M.build(pos, strategy)
   local runspec_strategy = nil
   if strategy == "dap" then
     if options.get().dap_go_enabled then
-      runspec_strategy = M.get_dap_config(test_name)
-      M.setup_debugging(test_folder_absolute_path)
+      runspec_strategy = dap.get_dap_config(test_name)
+      dap.setup_debugging(test_folder_absolute_path)
     end
   end
 
@@ -84,37 +85,6 @@ function M.build_runspec(
   end
 
   return run_spec
-end
-
-function M.setup_debugging(cwd)
-  local dap_go_opts = options.get().dap_go_opts or {}
-  local dap_go_opts_original = vim.deepcopy(dap_go_opts)
-  if dap_go_opts.delve == nil then
-    dap_go_opts.delve = {}
-  end
-  dap_go_opts.delve.cwd = cwd
-  require("dap-go").setup(dap_go_opts)
-
-  -- reset nvim-dap-go (and cwd) after debugging with nvim-dap
-  require("dap").listeners.after.event_terminated["neotest-golang-debug"] = function()
-    require("dap-go").setup(dap_go_opts_original)
-  end
-end
-
---- @param test_name string
---- @return table | nil
-function M.get_dap_config(test_name)
-  -- :help dap-configuration
-  local dap_config = {
-    type = "go",
-    name = "Neotest-golang",
-    request = "launch",
-    mode = "test",
-    program = "${fileDirname}",
-    args = { "-test.run", "^" .. test_name .. "$" },
-  }
-
-  return dap_config
 end
 
 return M
